backend/requirements.txt
Flask
Flask-CORS
google-cloud-storage
python-dotenv
gunicorn